--- a/lib/arrayDiff.js
+++ b/lib/arrayDiff.js
@@ -1,48 +1,48 @@
-/**
- * Find Delta between two arrays.
- *
- * Supports array of elements and array of objects
- * */
+'use strict'
 
-function findDeltaWithinTwoArraysOfObjects(listOriginal, listNew, uniqueProp) {
-    var delta = {
-        added: [],
-        removed: [],
-        common: []
-    };
-    var _deltaMap = {};
-    var _objHolder = {};
-    listOriginal.forEach(function (item) {
-        _objHolder[uniqueProp && getItemValue(item, uniqueProp) || item] = item;
-        _deltaMap[uniqueProp && getItemValue(item, uniqueProp) || item] = 'remove';
-    });
-    listNew.forEach(function (item) {
-        _objHolder[uniqueProp && getItemValue(item, uniqueProp) || item] = item;
-        var d = _deltaMap[uniqueProp && getItemValue(item, uniqueProp) || item];
-        if (Boolean(d) && d == 'remove') {
-            _deltaMap[uniqueProp && getItemValue(item, uniqueProp) || item] = 'noChange';
-        } else {
-            _deltaMap[uniqueProp && getItemValue(item, uniqueProp) || item] = 'added';
-        }
-    });
-    Object.keys(_deltaMap).forEach(function (id) {
-        if (_deltaMap[id] == 'remove') delta.removed.push(_objHolder[id]);
-        else if (_deltaMap[id] == 'added') delta.added.push(_objHolder[id]);
-        else if (_deltaMap[id] == 'noChange') delta.common.push(_objHolder[id]);
-    });
-    return delta;
+var FLAGS = {
+  REMOVED: '1',
+  COMMON: '0',
+  ADDED: '2'
 }
 
-function getItemValue(item, uniqueProp) {
-	if (typeof uniqueProp==='string'){
-		return item[uniqueProp];
-	}
-	else if	(typeof uniqueProp==='function')
-	{
-		return uniqueProp(item);
-	}
-	throw 'Property selector should either be a property name or selector function.';
+function arrayDiff (orig, dist, id) {
+  var _deltaMap = {}
+  var _objHolder = {}
+
+  orig.forEach(function (item) {
+    var key = get(item, id)
+    _objHolder[key] = item
+    _deltaMap[key] = FLAGS.REMOVED
+  })
+
+  dist.forEach(function (item) {
+    var key = get(item, id)
+    _objHolder[key] = item
+    var flag = _deltaMap[key]
+    if (flag === FLAGS.REMOVED) _deltaMap[key] = FLAGS.COMMON
+    else _deltaMap[key] = FLAGS.ADDED
+  })
+
+  var delta = { added: [], removed: [], common: [] }
+
+  Object.keys(_deltaMap).forEach(function (id) {
+    var value = _deltaMap[id]
+    var store
+    if (value === FLAGS.REMOVED) store = delta.removed
+    else if (value === FLAGS.ADDED) store = delta.added
+    else if (value === FLAGS.COMMON) store = delta.common
+    store.push(_objHolder[id])
+  })
+
+  return delta
 }
 
-//Lets export the method as module require
-exports = module.exports = findDeltaWithinTwoArraysOfObjects;
+function get (item, id) {
+  if (!id) return item
+  if (typeof id === 'string') return item[id]
+  if (typeof id === 'function') return id(item)
+  throw TypeError('Property selector should either be a property name or selector function.');
+}
+
+module.exports = arrayDiff