--- a/tests/test.js
+++ b/tests/test.js
@@ -3,59 +3,48 @@ var assert = require("assert"),
 
 describe('Array Diff Test Suite', function () {
 
-    before(function (done) {
-        this.timeout(10000);
-        done();
-    });
-
     describe('Element array diff Test', function () {
 
-        it('should differentiate between numerical elements', function (done) {
+        it('should differentiate between numerical elements', function () {
             var result = arrayDiff(
                 [1, 2, 3, 4, 5, 6],
                 [1, 2, 4, 5, 6, 0, 9, 10]
             );
-            console.log(result);
             assert(result);
             assert(result.added);
             assert(result.removed);
             assert.equal(result.added.length, 3);
             assert.equal(result.removed.length, 1);
-            done();
         });
 
-        it('should differentiate between string elements', function (done) {
+        it('should differentiate between string elements', function () {
             var result = arrayDiff(
                 ['a', 'b', 'c'],
                 ['c', 'd', 'e']
             );
-            console.log(result);
             assert(result);
             assert(result.added);
             assert(result.removed);
             assert.equal(result.added.length, 2);
             assert.equal(result.removed.length, 2);
-            done();
         });
 
-        it('should differentiate between mixed elements', function (done) {
+        it('should differentiate between mixed elements', function () {
             var result = arrayDiff(
                 ['a', 'b', 'c', 1, 3, 5, true],
                 ['c', 'd', 'e', 7, 3, 4, 5, false, true]
             );
-            console.log(result);
             assert(result);
             assert(result.added);
             assert(result.removed);
             assert.equal(result.added.length, 5);
             assert.equal(result.removed.length, 3);
-            done();
         });
     });
 
     describe('Object array diff Test', function () {
 
-        it('should differentiate between object elements', function (done) {
+        it('should differentiate between object elements', function () {
             var result = arrayDiff(
                 [
                     {id:1, name: 'a'},
@@ -71,16 +60,22 @@ describe('Array Diff Test Suite', function () {
                 ],
                 'id'
             );
-            console.log(result);
+
             assert(result);
             assert(result.added);
-            assert(result.removed);
             assert.equal(result.added.length, 1);
+            assert.deepEqual(result.added, [ { id: 7, name: 'e' } ])
+
+            assert(result.removed);
             assert.equal(result.removed.length, 3);
-            done();
+            assert.deepEqual(result.removed, [
+                {id:3, name: 'c'},
+                {id:4, name: 'd'},
+                {id:5, name: 'e'}
+            ])
         });
-		
-		 it('should differentiate between object elements when selector is a function', function (done) {
+
+        it('should differentiate between object elements when selector is a function', function () {
             var result = arrayDiff(
                 [
                     {id:1, name: 'a'},
@@ -95,20 +90,18 @@ describe('Array Diff Test Suite', function () {
                     {id:7, name: 'e'}
                 ],
                 function (x)
-				{
-					return x.id;
-				}
+          {
+              return x.id;
+          }
             );
-            console.log(result);
             assert(result);
             assert(result.added);
             assert(result.removed);
             assert.equal(result.added.length, 1);
             assert.equal(result.removed.length, 3);
-            done();
         });
-		
-		it('should differentiate between object elements when selector is a function regardless of object depth', function (done) {
+
+       it('should differentiate between object elements when selector is a function regardless of object depth', function () {
             var result = arrayDiff(
                   [
                     { id: 1, depthObj: { id: 1 }, name: 'a' },
@@ -126,20 +119,11 @@ describe('Array Diff Test Suite', function () {
                     return x.depthObj.id;
                 }
             );
-            console.log(result);
             assert(result);
             assert(result.added);
             assert(result.removed);
             assert.equal(result.added.length, 1);
             assert.equal(result.removed.length, 3);
-            done();
         });
-		
-		
-
-    });
-
-    after(function () {
-
     });
 });