Hello,

I needed a library like that. At first I looked in the npm registry, I didn't see a convenience library and I started creating my own solution. Later I  created a benchmark between your solution and mine and It said that your perfomance is better than my solution.

Looks like need less effort do twice `forEach` than find index of each key of one of the arrays in the another array.

After that I decided improve a little your library, like:

- Using `use strict` for better perfomance.
- Extract flags into constants.
- Minimize the size of the flags (1 char).
- Caching object keys values (this speed the process a lot).

Before my changes:

```
❯ node bench.js
simpleArrayDiff*1000: 131.993ms
hyperDiff*1000: 746.399ms
simpleArrayDiff*1000: 107.656ms
hyperDiff*1000: 727.353ms
```

After my changes:

```
❯ node bench.js
simpleArrayDiff*1000: 69.364ms
hyperDiff*1000: 811.103ms
simpleArrayDiff*1000: 69.525ms
hyperDiff*1000: 768.283ms
```

So that's my PR, I hope that you like it 👍 

That's the code of my library that I didn't publish because it looks worst: https://gist.github.com/Kikobeats/ff5dea8601462c3380277bfa4ba93a86

I undestand that find index in array have a costs, but still don't understand twice forEach looks better. If you can explain me, you're welcome 😄