@Kikobeats 2017-01-03
Published under https://github.com/Kikobeats/hyperdiff